spunoff input & output
